requirements.txt
ktrain==0.28.3
matplotlib==3.3.4
nltk==3.6.1
numpy==1.20.1
pandas==1.2.4
Pillow==9.0.0
plotly==5.3.1
seaborn==0.11.1
streamlit==1.4.0
tensorflow-cpu==2.7.0
transformers<=4.10.3
nltk
